max597x: Fix integer overflow

+++ drivers/regulator/max597x.c
@@ -227,7 +227,7 @@ static int max597x_set_ocp(struct regulator_dev *rdev, int lim_uA,
 		return -EINVAL;
 
 	/* Calc Vtrip threshold in uV. */
-	vthst = data->shunt_micro_ohms * lim_uA / 1000000;
+	vthst = (uint64_t)data->shunt_micro_ohms * (uint64_t)lim_uA / 1000000;
 
 	/* Calc fast Vtrip threshold in uV */
 	vthfst = vthst * (MAX5970_FAST2SLOW_RATIO / 100);
